update is now only one button, rewrote, removed, and moved some front-end and back-end functions
