+++ public/js/collection/sharedDataTable.js
@@ -19,7 +19,6 @@ function Initialize_table(){
         $table.scroller.toPosition(0);
     });
 }
-
 /**
  * Creates a datatable then Fills the table with JSON data
  * Link to Library: Datatables.net
@@ -32,7 +31,7 @@ function fill_table(data){
         data : data,
         dom : 'Bfrtip',
         buttons : button_builder(),
-        rowId : `symbol`,
+        rowId : `stock_id`,
         select : { selector: 'td:first-child', style : 'multi' },
         columns : column_builder(),
         fixedColumns: { leftColumns: 2 },
@@ -52,13 +51,9 @@ function fill_table(data){
 function button_builder(){
     let buttons = [
         'selectAll', 'selectNone',
-        {text: '<span class="fas fa-plus"></span> Add', className:"btn-sm", action: function(){add();}},
-        {text: '<span class="fas fa-trash-alt"></span> Delete', className:"btn-sm", action: function(){remove();}},
-        {text: '<span class="fas fa-sync-alt"></span> Refresh', className: "btn-sm", extend: 'collection',
-        buttons:[
-            {text: 'Prices', className:"btn-sm", action: function(){update('update_prices');}},
-            {text: 'Financials', className:"btn-sm", action: function(){update('update_financials');}},
-        ]},
+        {text: '<span class="fas fa-plus"></span> Add', className:"btn-sm", action: function(){add('/append?table=shared');}},
+        {text: '<span class="fas fa-trash-alt"></span> Delete', className:"btn-sm", action: function(){remove('/remove?table=shared');}},
+        {text: '<span class="fas fa-sync-alt"></span> Refresh', className: "btn-sm", action: function(){update('update_financials?table=shared');}},
         {text: `<span class="fas fa-calculator"></span> DCF`, className: "btn-sm", action: function(){calc_edit();}},
        
         {text: '<span class="fas fa-eye"></span> Show Selected', className:"btn-sm", action: function(){show_selected();}},
