+++ views/partials/datatable.hbs
@@ -122,11 +122,14 @@
 
 {{#if sc}}
     <script src="/js/collection/sharedDataTable.js"></script>
-    <script src="/js/collection/sharedButtons.js"></script>
     <script src="/js/collection/sharedColumns.js"></script>
     {{else}}
     <script src="/js/collection/privateDataTable.js"></script>
-    <script src="/js/collection/privateButtons.js"></script>
     <script src="/js/collection/privateColumns.js"></script>
 {{/if}}
+<script src="/js/collection/privateButtons.js"></script>
+<script src="/js/collection/modules/edit.js"></script>
+<script src="/js/collection/modules/dcfCalculations.js"></script>
+<script src="/js/collection/modules/aggregations.js"></script>
 <script src="/js/collection/common_functions.js"></script>
+
